+++ src/components/organisms/HomeworksDashboardTeacher.unit.js
@@ -11,10 +11,35 @@ describe("@components/organisms/HomeworksDashboardTeacher", () => {
 	const mockStore = {
 		homeworks: {
 			getters: {
-				getLoading: () => false,
+				getStatus: () => "completed",
+				isListFilled: () => true,
 				getOverDueHomeworks: () => overDueHomeworksTeacher,
 				getOpenHomeworksWithDueDate: () => dueDateHomeworksTeacher,
 				getOpenHomeworksWithoutDueDate: () => noDueDateHomeworksTeacher,
+			},
+		},
+	};
+
+	const mockStoreOnlyWithoutDueDate = {
+		homeworks: {
+			getters: {
+				getStatus: () => "completed",
+				isListFilled: () => true,
+				getOverDueHomeworks: () => [],
+				getOpenHomeworksWithDueDate: () => [],
+				getOpenHomeworksWithoutDueDate: () => noDueDateHomeworksTeacher,
+			},
+		},
+	};
+
+	const mockStoreOnlyWithDueDate = {
+		homeworks: {
+			getters: {
+				getStatus: () => "completed",
+				isListFilled: () => true,
+				getOverDueHomeworks: () => overDueHomeworksTeacher,
+				getOpenHomeworksWithDueDate: () => dueDateHomeworksTeacher,
+				getOpenHomeworksWithoutDueDate: () => [],
 			},
 		},
 	};
@@ -27,7 +52,7 @@ describe("@components/organisms/HomeworksDashboardTeacher", () => {
 
 	it(...isValidComponent(HomeworksDashboardTeacher));
 
-	it("Should render homeworks list component", () => {
+	it("Should render homeworks list component, with both panels expanded per default", () => {
 		const wrapper = mount(HomeworksDashboardTeacher, {
 			...createComponentMocks({
 				i18n: true,
@@ -37,6 +62,59 @@ describe("@components/organisms/HomeworksDashboardTeacher", () => {
 			vuetify,
 		});
 
+		const expansionPanels = wrapper.findAll(".v-expansion-panel");
+
 		expect(wrapper.findComponent(HomeworksList).exists()).toBe(true);
+		expect(expansionPanels.exists()).toBe(true);
+		expect(expansionPanels.at(0).classes()).not.toContain(
+			"v-expansion-panel--active"
+		);
+		expect(expansionPanels.at(1).classes()).toContain(
+			"v-expansion-panel--active"
+		);
+	});
+
+	it("Should render not disabled 'no due date' panel, and disabled empty other panel", async () => {
+		const wrapper = mount(HomeworksDashboardTeacher, {
+			...createComponentMocks({
+				i18n: true,
+				vuetify: true,
+				store: mockStoreOnlyWithoutDueDate,
+			}),
+			vuetify,
+		});
+
+		const expansionPanels = wrapper.findAll(".v-expansion-panel");
+
+		expect(expansionPanels.exists()).toBe(true);
+		expect(expansionPanels).toHaveLength(2);
+		expect(expansionPanels.at(0).classes()).not.toContain(
+			"v-expansion-panel--disabled"
+		);
+		expect(expansionPanels.at(1).classes()).toContain(
+			"v-expansion-panel--disabled"
+		);
+	});
+
+	it("Should render not disabled 'with due date' panel, and disabled empty other panel", async () => {
+		const wrapper = mount(HomeworksDashboardTeacher, {
+			...createComponentMocks({
+				i18n: true,
+				vuetify: true,
+				store: mockStoreOnlyWithDueDate,
+			}),
+			vuetify,
+		});
+
+		const expansionPanels = wrapper.findAll(".v-expansion-panel");
+
+		expect(expansionPanels.exists()).toBe(true);
+		expect(expansionPanels).toHaveLength(2);
+		expect(expansionPanels.at(0).classes()).toContain(
+			"v-expansion-panel--disabled"
+		);
+		expect(expansionPanels.at(1).classes()).not.toContain(
+			"v-expansion-panel--disabled"
+		);
 	});
 });