+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 - SC-9159 - Redirect the routes for news for the new v3 API
 - SC-9108 - Refactor news/create page and form component
+- SC-9189 - Adding expansion panels to teacher's task dashboard
 
 ### Added